+++ inc/server/bases/CommandManager.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   CommandManager.hpp                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kbutor-b <kbutor-b@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ysabik <ysabik@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 07:40:31 by ysabik            #+#    #+#             */
-/*   Updated: 2024/05/26 14:48:42 by kbutor-b         ###   ########.fr       */
+/*   Updated: 2024/05/28 03:15:37 by ysabik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,6 @@
 # include "QuitCommand.hpp"
 # include "TopicCommand.hpp"
 # include "UserCommand.hpp"
-# include "UsersCommand.hpp"
 
 class CommandManager {
 	private: