+++ src/server/main.cpp
@@ -6,7 +6,7 @@
 /*   By: ysabik <ysabik@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/04 02:53:43 by ysabik            #+#    #+#             */
-/*   Updated: 2024/05/28 02:24:36 by ysabik           ###   ########.fr       */
+/*   Updated: 2024/05/28 03:15:08 by ysabik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,7 +80,6 @@ int main(int argc, char **argv) {
 		.addCommand(new PongCommand("PONG", List<std::string>("PO")))
 		.addCommand(new PrivCommand("PRIVMSG", List<std::string>("M")("MSG")))
 		.addCommand(new UserCommand("USER", List<std::string>("U")))
-		.addCommand(new UsersCommand("USERS", List<std::string>("US")))
 		.addCommand(new TopicCommand("TOPIC", List<std::string>("T")))
 		.addCommand(new QuitCommand("QUIT", List<std::string>("Q")))
 		;